evolution(menu) : Deplace le Menu dans la classe PlayerSystem
Ajoute diverses ameliorations
