+++ Systems/PlayerSystem/Player.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using NWN.Enums.VisualEffect;
+using System;
 
 namespace NWN.Systems
 {
@@ -7,10 +8,14 @@ namespace NWN.Systems
     public class Player
     {
       public readonly uint oid;
+      public Menu menu { get; }
+
+      private uint blockingBoulder;
 
       public Player(uint oid)
       {
         this.oid = oid;
+        this.menu = new PrivateMenu(this);
       }
 
       public void EmitKeydown(KeydownEventArgs e)
@@ -28,6 +33,24 @@ namespace NWN.Systems
         {
           this.key = key;
         }
+      }
+
+      public void BoulderBlock ()
+      {
+        BoulderUnblock();
+        var location = NWScript.GetLocation(oid);
+        blockingBoulder = NWScript.CreateObject(Enums.ObjectType.Placeable, "plc_boulder", location, false);
+        NWNX.Object.SetPosition(oid, NWScript.GetPositionFromLocation(location));
+        NWScript.ApplyEffectToObject(
+          Enums.DurationType.Permanent,
+          NWScript.EffectVisualEffect((VisualEffect)Temporary.CutsceneInvisibility),
+          blockingBoulder
+        );
+      }
+
+      public void BoulderUnblock ()
+      {
+        NWScript.DestroyObject(blockingBoulder);
       }
     }
   }
